Refactor inbox utility imports for better organization
