+++ package.json
@@ -14,7 +14,7 @@
     "build": "tsc",
     "clean": "rimraf .data/ ||: && rimraf logs/ ||  :",
     "cli": "tsx scripts/cli.ts",
-    "count-duplicates": "NODE_OPTIONS='--experimental-specifier-resolution=node' tsx inboxes/count-duplicates.ts",
+    "count-duplicates": "NODE_OPTIONS='--experimental-specifier-resolution=node' tsx inboxes/utils.ts",
     "forked": "./suites/group/run.sh",
     "format": "prettier -w .",
     "functional": "yarn test suites/functional",
