+++ suites/large-groups/README.md
@@ -44,6 +44,42 @@
 | 140        | 20          | 2746        | -54  | 2800      | 536.00                 | 1244.64           | 0.20                  |
 | 160        | 20          | 3146        | -54  | 3200      | 541.00                 | 1449.46           | 0.17                  |
 
+## ~2000 Large Groups Performance Results
+
+| Group Size | Inst/Member | Actual Inst | Diff | Est. Inst | Add Members (ms) | SyncAll (ms) | Time per Install (ms) |
+| ---------- | ----------- | ----------- | ---- | --------- | ---------------- | ------------ | --------------------- |
+| 90         | 20          | 1743        | -57  | 1800      | 447.00           | 1001.52      | 0.26                  |
+| 120        | 15          | 1758        | -42  | 1800      | 332.00           | 775.86       | 0.19                  |
+| 180        | 10          | 1773        | -27  | 1800      | 393.00           | 851.00       | 0.22                  |
+| 190        | 10          | 1873        | -27  | 1900      | 316.00           | 944.03       | 0.17                  |
+| 130        | 15          | 1908        | -42  | 1950      | 404.00           | 1094.99      | 0.21                  |
+| 80         | 25          | 1928        | -72  | 2000      | 441.00           | 1099.99      | 0.23                  |
+| 100        | 20          | 1943        | -57  | 2000      | 510.00           | 861.25       | 0.26                  |
+| 200        | 10          | 1973        | -27  | 2000      | 364.00           | 994.03       | 0.18                  |
+| 140        | 15          | 2058        | -42  | 2100      | 429.00           | 1241.45      | 0.21                  |
+| 110        | 20          | 2144        | -56  | 2200      | 444.00           | 1270.69      | 0.21                  |
+
+## Large Groups Performance Results
+
+| Group Size | Inst/Member | Actual Inst | Diff | Est. Inst | Add Members (ms) | SyncAll (ms) | Time per Install (ms) |
+| ---------- | ----------- | ----------- | ---- | --------- | ---------------- | ------------ | --------------------- |
+| 60         | 30          | 1713        | -87  | 1800      | 422.00           | 997.53       | 0.25                  |
+| 90         | 20          | 1743        | -57  | 1800      | 372.00           | 747.72       | 0.21                  |
+| 120        | 15          | 1758        | -42  | 1800      | 401.00           | 1007.16      | 0.23                  |
+| 180        | 10          | 1774        | -26  | 1800      | 443.00           | 1138.13      | 0.25                  |
+| 190        | 10          | 1873        | -27  | 1900      | 441.00           | 1209.25      | 0.24                  |
+| 130        | 15          | 1908        | -42  | 1950      | 403.00           | 1122.18      | 0.21                  |
+| 80         | 25          | 1928        | -72  | 2000      | 543.00           | 1076.16      | 0.28                  |
+| 100        | 20          | 1944        | -56  | 2000      | 351.00           | 1141.92      | 0.18                  |
+| 200        | 10          | 1974        | -26  | 2000      | 449.00           | 1209.01      | 0.23                  |
+| 70         | 30          | 2013        | -87  | 2100      | 431.00           | 1123.54      | 0.21                  |
+| 140        | 15          | 2058        | -42  | 2100      | 435.00           | 1009.45      | 0.21                  |
+| 110        | 20          | 2143        | -57  | 2200      | 416.00           | 1257.02      | 0.19                  |
+
+---
+
+## Insights
+
 A 3200 installations group gives:
 
 ```
@@ -58,18 +94,3 @@ Networks breaks after this test with:
 ❌ Error updating inbox 0x41c4d0e3736667e9993e384c7c85e0b90771d7c1:
 [2025-06-09T18:09:29.827Z] [error] ERROR api client error api client at endpoint "get_inbox_ids" has error status: Unknown, message: "failed to connect to `host=mlsdb user=postgres database=postgres`: hostname resolving error (lookup mlsdb on 127.0.0.11:53: no such host)", details: [], metadata: MetadataMap { headers: {"content-type": "application/grpc"} }
 ```
-
-## Large Groups Performance Results
-
-| Group Size | Inst/Member | Actual Inst | Diff | Est. Inst | Add Members (ms) | SyncAll (ms) | Time per Install (ms) |
-| ---------- | ----------- | ----------- | ---- | --------- | ---------------- | ------------ | --------------------- |
-| 90         | 20          | 1743        | -57  | 1800      | 447.00           | 1001.52      | 0.26                  |
-| 120        | 15          | 1758        | -42  | 1800      | 332.00           | 775.86       | 0.19                  |
-| 180        | 10          | 1773        | -27  | 1800      | 393.00           | 851.00       | 0.22                  |
-| 190        | 10          | 1873        | -27  | 1900      | 316.00           | 944.03       | 0.17                  |
-| 130        | 15          | 1908        | -42  | 1950      | 404.00           | 1094.99      | 0.21                  |
-| 80         | 25          | 1928        | -72  | 2000      | 441.00           | 1099.99      | 0.23                  |
-| 100        | 20          | 1943        | -57  | 2000      | 510.00           | 861.25       | 0.26                  |
-| 200        | 10          | 1973        | -27  | 2000      | 364.00           | 994.03       | 0.18                  |
-| 140        | 15          | 2058        | -42  | 2100      | 429.00           | 1241.45      | 0.21                  |
-| 110        | 20          | 2144        | -56  | 2200      | 444.00           | 1270.69      | 0.21                  |
